separate logout widget and add consumer for light theme
